src/main/java/Card.java
import org.example.Rank;
import org.example.Suit;

public class Card {
    private Suit suit;
    private Rank rank;

    public Card() {
    }

    public Card(Suit suit, Rank rank) {
        this.suit = suit;
        this.rank = rank;
    }

    public Suit getSuit() {
        return suit;
    }

    public void setSuit(Suit suit) {
        this.suit = suit;
    }

    public Rank getRank() {
        return rank;
    }

    public void setRank(Rank rank) {
        this.rank = rank;
    }

    @Override
    public String toString() {
        return "Card{" +
                "suit='" + suit + '\'' +
                ", rank='" + rank + '\'' +
                '}';
    }

    public static void main(String[] args) {
        Card card1 = new Card(Suit.CLUBS, Rank.JACK);
        Card card2 = new Card(Suit.DIAMONDS, Rank.TWO);
        System.out.println();

    }

    public int getValue() {
        return switch (this.rank) {
            case JACK, QUEEN, KING -> 10;
            default -> this.rank.ordinal() + 1;
        };
    }
}

